Improve struct packing, reduced floating point precision from long double to double for better memory efficiency, will be improved with scientific notation

+++ src/parsing/parserUtils.c
@@ -110,7 +110,7 @@ ASTNode *createASTNode(Token *token) {
     case TOKEN_NUMBER:
         node->type = NODE_DOUBLE;
         char *end;
-        long double value = strtod(token->value, &end);
+        double value = strtod(token->value, &end);
         
         if (value == (int) value) {
             node->type = NODE_INTEGER;
